+++ geotrek/common/mixins/views.py
@@ -268,7 +268,7 @@ class BelongStructureMixin:
                     "Access is restricted because not all selected items belong to your structure. Use the structure filter to select only authorized items."
                 ),
             )
-            return HttpResponseRedirect(self.get_redirect_url())
+            return HttpResponseRedirect(self.get_success_url())
 
         return response