+++ geotrek/signage/views.py
@@ -350,7 +350,7 @@ class BladeMultiDelete(MapEntityMultiDelete):
                     "Access is restricted because not all selected items belong to your structure. Use the structure filter to select only authorized items."
                 ),
             )
-            return HttpResponseRedirect(self.get_redirect_url())
+            return HttpResponseRedirect(self.get_success_url())
 
         return response
 
@@ -362,6 +362,24 @@ class BladeMultiUpdate(MapEntityMultiUpdate):
         editable_fields = super().get_editable_fields()
         editable_fields.remove("topology")
         return editable_fields
+
+    def form_valid(self, form):
+        signage = form.cleaned_data.get("signage", None)
+        user = self.request.user
+        user_structure = user.profile.structure
+
+        if (
+            not user.has_perm("authent.can_bypass_structure")
+            and signage
+            and signage != "unknown"
+            and signage.structure != user_structure
+        ):
+            messages.warning(
+                self.request,
+                _("Selected signage is not in your structure"),
+            )
+            return self.form_invalid(form)
+        return super().form_valid(form)
 
     def get(self, request, *args, **kwargs):
         # check pks definition first to avoid get_queryset error
@@ -389,6 +407,6 @@ class BladeMultiUpdate(MapEntityMultiUpdate):
                     "Access is restricted because not all selected items belong to your structure. Use the structure filter to select only authorized items."
                 ),
             )
-            return HttpResponseRedirect(self.get_redirect_url())
+            return HttpResponseRedirect(self.get_success_url())
 
         return response